+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-   Fixed issue generating Sharpire exes
+
 ## [6.0.0] - 2025-03-25
 
 ### Highlights